automation/telegram_bot/requirements.txt
aiogram>=3.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pywin32>=306
aiofiles>=23.0.0
